+++ test/fixtures/rdf_mapper_annotations/examples/provides.rdf_mapper.g.dart
@@ -82,7 +82,7 @@ class ChildMapper implements GlobalResourceMapper<Child> {
 /// Generated mapper for [String] global resources.
 ///
 /// This mapper handles serialization and deserialization between Dart objects
-/// and RDF terms for iri terms of type String.
+/// and RDF terms for iri terms of type `String`.
 class ParentSiblingIdMapper implements IriTermMapper<String> {
   static final RegExp _regex = RegExp(
     r'^(?<baseUri>.*)/(?<parentId>[^/]*)/sibling/(?<siblingId>[^/]*)\.ttl$',